Move details and description to dedicated component
